chore: Fix TopRatedRepositoryImplTest and remove duplicates
